feat: change lat lon to float

+++ database/migrations/1688542318500_addresses.ts
@@ -13,8 +13,8 @@ export default class extends BaseSchema {
       table.string('city').notNullable()
       table.string('country').notNullable()
 
-      table.decimal('lat').notNullable()
-      table.decimal('lon').notNullable()
+      table.float('lat').notNullable()
+      table.float('lon').notNullable()
 
       table.string('user_id').notNullable()
     })
